# api/swagger.yaml
openapi: 3.0.3
info:
  title: Athlete Hub API
  description: |
    REST API for the Athlete Hub sports events platform.

    ## Authentication
    - **Protected routes** require `Authorization: Bearer <session_token>`
    - **Public routes** require no authentication

    ## Base URL
    All endpoints are prefixed with `/api`.
  version: "1.0.0"
  contact:
    name: Athlete Hub Dev
servers:
  - url: /api
    description: Same-origin (development & production)
...
paths:
...
  /athlete/registrations/{publicUuid}/transfer:
    post:
      tags: [Athlete]
      summary: Transfer registration to another athlete by email
      security:
        - bearerAuth: []
      parameters:
        - name: publicUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipientEmail]
              properties:
                recipientEmail:
                  type: string
                  format: email
      responses:
        "200":
          description: Transfer completed
        "400":
          description: Transfers not allowed or invalid state
        "403":
          description: Not the registration owner
        "404":
          description: Registration or recipient not found
        "409":
          description: Recipient already registered
...
  /organizer/events/{eventId}/messages/bulk:
    post:
      tags: [Organizer]
      summary: "[Deprecated] Queue bulk email"
      deprecated: true
      description: Returns 410. Use `/organizer/events/{eventId}/broadcasts` (Resend Broadcast) instead.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, body]
              properties:
                subject:
                  type: string
                body:
                  type: string
      responses:
        "410":
          description: Deprecated — use event broadcasts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...